moved to artist manager
