+++ Bess/CMakeLists.txt
@@ -64,7 +64,10 @@ set(Header_Files
 "include/scene/renderer/renderer.h"
 "include/scene/scene.h"
 "include/scene/scene_serializer.h"
-"include/scene/artist.h"
+"include/scene/artist/base_artist.h"
+"include/scene/artist/schematic_artist.h"
+"include/scene/artist/nodes_artist.h"
+"include/scene/artist/artist_manager.h"
 "include/scene/components/components.h"
 "include/scene/components/json_converters.h"
 "include/scene/components/non_sim_comp.h"
@@ -126,7 +129,10 @@ set(Source_Files
 "src/scene/components/non_sim_comp.cpp"
 "src/scene/scene.cpp"
 "src/scene/scene_serializer.cpp"
-"src/scene/artist.cpp"
+"src/scene/artist/base_artist.cpp"
+"src/scene/artist/schematic_artist.cpp"
+"src/scene/artist/nodes_artist.cpp"
+"src/scene/artist/artist_manager.cpp"
 "src/scene/renderer/font.cpp"
 "src/scene/renderer/msdf_font.cpp"
 "src/scene/renderer/gl/gl_wrapper.cpp"
