--- Bess/include/scene/artist.h
@@ -1,85 +1,0 @@
-#pragma once
-
-#include "bess_uuid.h"
-#include "entt/entity/fwd.hpp"
-#include "ext/vector_float3.hpp"
-#include "scene/components/components.h"
-#include "scene/scene.h"
-#include "types.h"
-
-#include <memory.h>
-
-namespace Bess::Canvas {
-
-    struct ArtistCompSchematicInfo {
-        float inpConnStart = 0.f;
-        float outConnStart = 0.f;
-        float inpPinStart = 0.f;
-        float outPinStart = 0.f;
-        float rb = 0.f;
-        float width, height = 0.f;
-        bool shouldDraw = true;
-    };
-
-    struct ArtistTools {
-        std::array<std::shared_ptr<Gl::SubTexture>, 8> sevenSegDispTexs;
-    };
-
-    struct ArtistInstructions {
-        bool isSchematicView = false;
-    };
-
-    class Artist {
-      public:
-        static void init();
-
-        static Scene *sceneRef;
-        static void drawSimEntity(entt::entity entity);
-        static void drawSimEntity(
-            entt::entity entity,
-            Components::TagComponent &tagComp,
-            Components::TransformComponent &transform,
-            Components::SpriteComponent &spriteComp,
-            Components::SimulationComponent &simComponent);
-
-        static void drawHeaderLessComp(entt::entity entity,
-                                       Components::TagComponent &tagComp,
-                                       Components::TransformComponent &transform,
-                                       Components::SpriteComponent &spriteComp,
-
-                                       Components::SimulationComponent &simComp);
-        static void drawNonSimEntity(entt::entity entity);
-
-        static void drawConnectionEntity(entt::entity entity);
-        static glm::vec3 getSlotPos(const Components::SlotComponent &comp,
-                                    const Components::TransformComponent &parentTransform);
-        // use in schematic mode
-        static glm::vec3 getPinPos(const Components::SlotComponent &comp);
-        static glm::vec3 getPinPos(UUID uuid);
-        static void drawGhostConnection(const entt::entity &startEntity, const glm::vec2 pos);
-        static void drawConnection(const UUID &id, entt::entity inputEntity, entt::entity outputEntity, bool isSelected);
-
-      public:
-        static void setInstructions(const ArtistInstructions &value);
-
-      private:
-        static void paintSchematicView(entt::entity entity);
-
-        static void paintSlot(uint64_t id, uint64_t parentId, const glm::vec3 &pos, float angle,
-                              const std::string &label, float labelDx, SimEngine::LogicState state, bool isConnected, SimEngine::ExtendedPinType extendedType);
-
-        static void drawSlots(const Components::SimulationComponent &comp, const Components::TransformComponent &transformComp);
-
-        static void drawSevenSegDisplay(entt::entity entity,
-                                        Components::TagComponent &tagComp,
-                                        Components::TransformComponent &transform,
-                                        Components::SpriteComponent &spriteComp,
-                                        Components::SimulationComponent &simComponent);
-
-        static ArtistCompSchematicInfo getCompSchematicInfo(entt::entity ent);
-
-        static ArtistTools m_artistTools;
-
-        static ArtistInstructions m_instructions;
-    };
-} // namespace Bess::Canvas
