+++ Bess/src/scene/artist/nodes_artist.cpp
@@ -1,0 +1,271 @@
+#include "scene/artist/nodes_artist.h"
+#include "component_catalog.h"
+#include "component_types/component_types.h"
+#include "entt/entity/fwd.hpp"
+#include "ext/vector_float3.hpp"
+#include "scene/components/components.h"
+#include "scene/renderer/renderer.h"
+#include "scene/scene.h"
+#include "settings/viewport_theme.h"
+#include "simulation_engine.h"
+#include "types.h"
+#include <cstdint>
+#include <string>
+
+using Renderer = Bess::Renderer2D::Renderer;
+
+namespace Bess::Canvas {
+
+    NodesArtist::NodesArtist(Scene *scene) : BaseArtist(scene) {
+    }
+
+    void NodesArtist::drawSimEntity(entt::entity entity) {
+        auto &registry = m_sceneRef->getEnttRegistry();
+        auto &tagComp = registry.get<Components::TagComponent>(entity);
+        auto &transform = registry.get<Components::TransformComponent>(entity);
+        auto &spriteComp = registry.get<Components::SpriteComponent>(entity);
+        auto &simComp = registry.get<Components::SimulationComponent>(entity);
+
+        drawSimEntity(entity, tagComp, transform, spriteComp, simComp);
+    }
+
+    void NodesArtist::drawSimEntity(
+        entt::entity entity,
+        Components::TagComponent &tagComp,
+        Components::TransformComponent &transform,
+        Components::SpriteComponent &spriteComp,
+        Components::SimulationComponent &simComponent) {
+
+        auto &registry = m_sceneRef->getEnttRegistry();
+
+        if (tagComp.isSimComponent && tagComp.type.simCompType == SimEngine::ComponentType::SEVEN_SEG_DISPLAY) {
+            drawSevenSegDisplay(entity, tagComp, transform, spriteComp, simComponent);
+            return;
+        }
+
+        // Handle headerless components (INPUT, OUTPUT, STATE_MONITOR) in nodes view
+        if (tagComp.type.simCompType == SimEngine::ComponentType::INPUT ||
+            tagComp.type.simCompType == SimEngine::ComponentType::OUTPUT ||
+            tagComp.type.simCompType == SimEngine::ComponentType::STATE_MONITOR) {
+            drawHeaderLessComp(entity, tagComp, transform, spriteComp, simComponent);
+            return;
+        }
+
+        auto pos = transform.position;
+        auto rotation = transform.angle;
+        auto scale = transform.scale;
+        int maxRows = std::max(simComponent.inputSlots.size(), simComponent.outputSlots.size());
+        scale.y = 18.f + 4.f + (maxRows * SLOT_ROW_SIZE); // headerHeight + rowGap + (maxRows * SLOT_ROW_SIZE)
+        transform.scale = scale;
+
+        float headerHeight = 18.f; // componentStyles.headerHeight
+        auto headerPos = glm::vec3(pos.x, pos.y - scale.y / 2.f + headerHeight / 2.f, pos.z);
+
+        bool isSelected = registry.any_of<Components::SelectedComponent>(entity);
+        auto border = isSelected ? ViewportTheme::colors.selectedComp : spriteComp.borderColor;
+
+        uint64_t id = (uint64_t)entity;
+
+        glm::vec3 textPos = glm::vec3(pos.x - scale.x / 2.f + 8.f, headerPos.y + 4.f, pos.z + 0.0005f); // paddingX, paddingY
+
+        Renderer2D::QuadRenderProperties props;
+        props.angle = rotation;
+        props.borderRadius = spriteComp.borderRadius;
+        props.borderSize = spriteComp.borderSize;
+        props.borderColor = border;
+        props.isMica = true;
+        props.hasShadow = true;
+
+        Renderer::quad(pos, glm::vec2(scale), spriteComp.color, id, props);
+
+        props = {};
+        props.angle = rotation;
+        props.borderSize = glm::vec4(0.f);
+        props.borderRadius = glm::vec4(0, 0, spriteComp.borderRadius.x - spriteComp.borderSize.x, spriteComp.borderRadius.y - spriteComp.borderSize.y);
+        props.isMica = true;
+
+        Renderer::quad(headerPos,
+                       glm::vec2(scale.x - spriteComp.borderSize.w - spriteComp.borderSize.y, headerHeight - spriteComp.borderSize.x - spriteComp.borderSize.z),
+                       spriteComp.headerColor,
+                       id,
+                       props);
+
+        Renderer::msdfText(tagComp.name, textPos, 10.f, ViewportTheme::colors.text, id, rotation); // headerFontSize
+
+        drawSlots(entity, simComponent, transform);
+    }
+
+    void NodesArtist::drawHeaderLessComp(entt::entity entity,
+                                         Components::TagComponent &tagComp,
+                                         Components::TransformComponent &transform,
+                                         Components::SpriteComponent &spriteComp,
+                                         Components::SimulationComponent &simComp) {
+        auto &registry = m_sceneRef->getEnttRegistry();
+
+        auto labelSize = Renderer::getMSDFTextRenderSize(tagComp.name, 10.f); // headerFontSize
+
+        uint64_t id = (uint64_t)entity;
+        auto pos = transform.position;
+        auto rotation = transform.angle;
+        auto scale = transform.scale;
+        float columnSize = SLOT_COLUMN_SIZE;
+
+        float xSize = labelSize.x;
+        xSize += 8.f * 2.f; // paddingX * 2
+
+        float labelXGapL = 0.f, labelXGapR = 0.f;
+        float labelLOffset = 8.f; // paddingX
+
+        if (!simComp.inputSlots.empty()) {
+            labelXGapL = 2.f;
+            xSize += SLOT_COLUMN_SIZE + labelXGapL;
+            labelLOffset += SLOT_COLUMN_SIZE + labelXGapL;
+        }
+
+        if (!simComp.outputSlots.empty()) {
+            labelXGapR = 2.f;
+            xSize += SLOT_COLUMN_SIZE + labelXGapR;
+        }
+
+        scale.x = xSize;
+
+        scale.y = SLOT_ROW_SIZE * std::max(simComp.inputSlots.size(), simComp.outputSlots.size());
+        transform.scale = scale;
+
+        bool isSelected = registry.any_of<Components::SelectedComponent>(entity);
+        auto border = isSelected ? ViewportTheme::colors.selectedComp : spriteComp.borderColor;
+
+        Renderer2D::QuadRenderProperties props;
+        props.borderRadius = spriteComp.borderRadius;
+        props.borderColor = border;
+        props.borderSize = spriteComp.borderSize;
+        props.isMica = true;
+        Renderer::quad(pos, glm::vec2(scale), spriteComp.color, id, props);
+
+        glm::vec3 textPos = glm::vec3(
+            pos.x - (scale.x / 2.f) + labelLOffset,
+            pos.y + (10.f / 2.f) - 1.f, pos.z + 0.0005f); // headerFontSize / 2
+
+        auto name = tagComp.name;
+        Renderer::msdfText(name, textPos, 10.f, ViewportTheme::colors.text, id); // headerFontSize
+
+        drawSlots(entity, simComp, transform);
+    }
+
+    glm::vec3 NodesArtist::getSlotPos(const Components::SlotComponent &comp, const Components::TransformComponent &parentTransform) {
+        auto pPos = parentTransform.position;
+        auto pScale = parentTransform.scale;
+
+        auto slotdx = SLOT_DX;
+
+        auto posX = pPos.x - pScale.x / 2.f;
+
+        bool isOutputSlot = comp.slotType == Components::SlotType::digitalOutput;
+        if (isOutputSlot) {
+            slotdx *= -1;
+            posX += pScale.x;
+        }
+
+        posX += slotdx;
+        float posY = pPos.y - pScale.y / 2.f + (SLOT_ROW_SIZE * comp.idx) + SLOT_ROW_SIZE / 2.f;
+
+        auto parentEntt = m_sceneRef->getEntityWithUuid(comp.parentId);
+        const auto &isNonHeader = m_sceneRef->getEnttRegistry().any_of<Components::SimulationInputComponent, Components::SimulationOutputComponent, Components::SimulationStateMonitor>(parentEntt);
+        if (!isNonHeader)
+            posY += SLOT_START_Y;
+
+        return glm::vec3(posX, posY, pPos.z + 0.0005);
+    }
+
+    void NodesArtist::paintSlot(uint64_t id, uint64_t parentId, const glm::vec3 &pos,
+                                float angle, const std::string &label, float labelDx,
+                                SimEngine::LogicState state, bool isConnected, SimEngine::ExtendedPinType extendedType) {
+        auto bg = ViewportTheme::colors.stateLow;
+        if (extendedType == SimEngine::ExtendedPinType::inputClock) {
+            if ((bool)state) {
+                bg = ViewportTheme::colors.clockConnectionHigh;
+            } else {
+                bg = ViewportTheme::colors.clockConnectionLow;
+            }
+        } else {
+            switch (state) {
+            case SimEngine::LogicState::low:
+                bg = ViewportTheme::colors.stateLow;
+                break;
+            case SimEngine::LogicState::high:
+                bg = ViewportTheme::colors.stateHigh;
+                break;
+            case SimEngine::LogicState::unknown:
+                bg = ViewportTheme::colors.stateUnknow;
+                break;
+            case SimEngine::LogicState::high_z:
+                bg = ViewportTheme::colors.stateHighZ;
+                break;
+            }
+        }
+
+        auto border = bg;
+        if (!isConnected)
+            bg.a = 0.1f;
+
+        float ir = componentStyles.slotRadius - componentStyles.slotBorderSize;
+        float r = componentStyles.slotRadius;
+
+        if (extendedType == SimEngine::ExtendedPinType::inputClear) {
+            Renderer2D::QuadRenderProperties props;
+            props.borderColor = border;
+            props.borderRadius = glm::vec4(2.5f);
+            props.borderSize = glm::vec4(componentStyles.slotBorderSize + 0.5);
+            Renderer::quad(pos, glm::vec2(r * 2.f), glm::vec4(0.f), id, props);
+            props.borderSize = {};
+            props.borderRadius = glm::vec4(1.5f);
+            Renderer::quad(pos, glm::vec2((ir - 1) * 2.f), glm::vec4(bg), id, props);
+        } else {
+            Renderer::circle(pos, r, border, id, ir);
+            Renderer::circle(pos, ir - 1.f, bg, id);
+        }
+
+        float labelX = pos.x + labelDx;
+        float dY = componentStyles.slotRadius - std::abs(componentStyles.slotRadius * 2.f - componentStyles.slotLabelSize) / 2.f;
+        Renderer::msdfText(label, {labelX, pos.y + dY, pos.z}, componentStyles.slotLabelSize, ViewportTheme::colors.text, parentId, angle);
+    }
+
+    void NodesArtist::drawSlots(const entt::entity parentEntt, const Components::SimulationComponent &comp, const Components::TransformComponent &transformComp) {
+        auto def = SimEngine::ComponentCatalog::instance().getComponentDefinition(comp.type);
+        auto &registry = m_sceneRef->getEnttRegistry();
+        auto slotsView = registry.view<Components::SlotComponent>();
+
+        float labeldx = componentStyles.slotMargin + (componentStyles.slotRadius * 2.f);
+
+        auto compState = SimEngine::SimulationEngine::instance().getComponentState(comp.simEngineEntity);
+
+        float angle = transformComp.angle;
+        auto [inpDetails, outDetails] = def->getPinDetails();
+
+        std::string label;
+        for (size_t i = 0; i < comp.inputSlots.size(); i++) {
+            auto slot = m_sceneRef->getEntityWithUuid(comp.inputSlots[i]);
+            auto state = compState.inputStates[i];
+            auto isConnected = compState.inputConnected[i];
+            auto &slotComp = slotsView.get<Components::SlotComponent>(slot);
+            auto slotPos = getSlotPos(slotComp, transformComp);
+            uint64_t parentId = (uint64_t)m_sceneRef->getEntityWithUuid(slotComp.parentId);
+            label = inpDetails.size() > i ? inpDetails[i].name : "X" + std::to_string(i);
+            paintSlot((uint64_t)slot, parentId, slotPos, angle, label, labeldx, state.state, isConnected,
+                      inpDetails.size() > i ? inpDetails[i].extendedType : SimEngine::ExtendedPinType::none);
+        }
+
+        for (size_t i = 0; i < comp.outputSlots.size(); i++) {
+            auto slot = m_sceneRef->getEntityWithUuid(comp.outputSlots[i]);
+            auto state = compState.outputStates[i];
+            auto isConnected = compState.outputConnected[i];
+            auto &slotComp = slotsView.get<Components::SlotComponent>(slot);
+            auto slotPos = getSlotPos(slotComp, transformComp);
+            uint64_t parentId = (uint64_t)m_sceneRef->getEntityWithUuid(slotComp.parentId);
+            label = outDetails.size() > i ? outDetails[i].name : "Y" + std::to_string(i);
+            float labelWidth = Renderer::getMSDFTextRenderSize(label, componentStyles.slotLabelSize).x;
+            paintSlot((uint64_t)slot, parentId, slotPos, angle, label, -labeldx - labelWidth, state.state, isConnected,
+                      outDetails.size() > i ? outDetails[i].extendedType : SimEngine::ExtendedPinType::none);
+        }
+    }
+} // namespace Bess::Canvas
