+++ Bess/src/ui/ui_main/ui_main.cpp
@@ -13,7 +13,7 @@
 
 #include "camera.h"
 #include "pages/main_page/main_page_state.h"
-#include "scene/artist.h"
+#include "scene/artist/artist_manager.h"
 #include "scene/renderer/gl/gl_wrapper.h"
 #include "scene/scene.h"
 #include "ui/icons/FontAwesomeIcons.h"
